+++ src/database/database.json
@@ -1,1 +1,1 @@
-[{"id":1,"name":"Sheila Melo","email":"sheila.melo@gmail.com","deleted":false},{"id":2,"name":"usuario_02","email":"usuario_2@gmail.com","deleted":false},{"id":3,"name":"usuario_03","email":"usuario_3@gmail.com","deleted":false},{"id":4,"name":"usuario_04","email":"usuario_3@gmail.com","deleted":true},{"id":5,"name":"usuario_05","email":"usuario_5@gmail.com","deleted":true},{"id":6,"name":"usuario_06","email":"usuario_6@gmail.com","deleted":true}]
+[{"id":1,"name":"Sheila Mello","email":"sheila.mello@zlpet.com.br","deleted":false},{"id":2,"name":"usuario_02","email":"usuario_2@gmail.com","deleted":false},{"id":3,"name":"usuario_03","email":"usuario_3@gmail.com","deleted":false},{"id":4,"name":"usuario_04","email":"usuario_3@gmail.com","deleted":true},{"id":5,"name":"usuario_05","email":"usuario_5@gmail.com","deleted":true},{"id":6,"name":"usuario_06","email":"usuario_6@gmail.com","deleted":true},{"id":7,"name":"Zoro","email":"zoro@zlpet.com.br","deleted":false},{"id":8,"name":"Teste","email":"teste@teste.com","deleted":false},{"id":9,"name":"Diego","email":"diego@uol.com","deleted":false},{"id":10,"name":"usuario01","email":"usuario01@empresa.com.br","deleted":false},{"id":11,"name":"usuario03","email":"usuario03@empresaa.com","deleted":false},{"id":12,"name":"usuario10","email":"usuario10@empresa_c.com.br","deleted":false},{"id":13,"name":"hamilton","email":"hamilton44@mercedes.f1.com","deleted":false},{"id":14,"name":"massa","email":"massa@italiana.com","deleted":false},{"id":15,"name":"abcd","email":"abcd@alfabeto.com.br","deleted":false},{"id":16,"name":"usuario15","email":"usuario15@empresa_a.com","deleted":false}]
